add: firebaseのロジック追加した「
